Added pipeline for Programs Service
ECOM-5775
